requirements.txt
# Core Framework
fastapi>=0.115.7
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Google ADK and AI
google-adk==1.16.0
google-generativeai==0.8.0
openai==1.40.0

# Image Generation
requests>=2.32.4
httpx>=0.27.0  # For product template API client (Phase 7.1.2)
pillow==10.4.0

# Async & Concurrency
aiohttp==3.10.0
asyncio==3.4.3

# Database
prisma==0.15.0

# Environment & Config
python-dotenv==1.0.1
pyyaml==6.0.2

# Monitoring & Logging
structlog==24.4.0
prometheus-client==0.20.0

# Security
cryptography==43.0.0
python-jose[cryptography]==3.3.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Retry & Resilience
tenacity==8.5.0

# Language Detection
langdetect==1.0.9

# Development
black==24.8.0
flake8==7.1.0
mypy==1.11.0
isort==5.13.0
